pyproject.toml
[project]
name = "cubot"
version = "0.1.0"
description = "Persuasive chatbot"
readme = "README.md"
requires-python = ">=3.12"

[dependency-groups]
main = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.0",
    "dotenv>=0.9.9",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "ipython>=9.4.0",
    "jupyter>=1.1.1",
    "notebook<7.0.0",
    "openai>=1.95.1",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
]

dev = [
    "black>=25.1.0",
    "django-extensions>=4.1",
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
]
